+++ src/api/index.tsx
@@ -1,8 +1,12 @@
-import { api as apiUtil, createQueryString } from '@/utils'
+import { api as apiUtil } from '@/utils'
 import {
   TGetScore,
-  TGetScoreResponse
+  TGetScoreResponse,
 } from './get-score'
+import {
+  TVerifyProofsResponse,
+  TVerifyProofs
+} from './verify-proofs'
 import configs from '../configs'
 
 const getScore: TGetScore = (
@@ -13,11 +17,35 @@ const getScore: TGetScore = (
       'GET',
       {
         'Authorization': `Bearer ${configs.ZUPLO_API_KEY}`,
+        'content-type': 'application/json',
+
       }
     )
 }
+
+const verifyProofs: TVerifyProofs = (
+  proofs,
+  chainId,
+  registryAddress
+) => {
+    return apiUtil<TVerifyProofsResponse>(
+      `${configs.ZUPLO_API_URL}/v1/proofs/verify`,
+      'POST',
+      {
+        'Authorization': `Bearer ${configs.ZUPLO_API_KEY}`,
+        'content-type': 'application/json'
+      },
+      {
+        proofs,
+        registry: registryAddress,
+        chain_id: chainId
+      }
+    )
+}
+
 const api = {
-  getScore
+  getScore,
+  verifyProofs
 }
 
 export default api