+++ src/modules/bring-id-sdk/types/constructor-args/index.tsx
@@ -1,0 +1,7 @@
+import { TMode } from "@/types"
+
+type TConstructorArgs = {
+  mode: TMode
+}
+
+export default TConstructorArgs