+++ src/modules/bring-id-sdk/types/index.tsx
@@ -1,9 +1,13 @@
-import TCheckExtensionStatus from './check-extension-status'
 import TVerifyHumanity from './verify-humanity'
 import TGetAddressScore from './get-address-score'
+import TDestroy from './destroy'
+import TVerifyProofs from './verify-proofs'
+import TConstructorArgs from './constructor-args'
 
 export {
-  TCheckExtensionStatus,
   TVerifyHumanity,
-  TGetAddressScore
+  TGetAddressScore,
+  TDestroy,
+  TVerifyProofs,
+  TConstructorArgs
 }