+++ src/modules/bring-id-sdk/types/verify-proofs/index.tsx
@@ -1,0 +1,11 @@
+import { TSemaphoreProof } from "@/types"
+import { ethers } from 'ethers'
+
+type TArgs = {
+  proofs: TSemaphoreProof[]
+  provider?: ethers.JsonRpcProvider
+}
+
+type TVerifyProofs = (payload: TArgs) => Promise<boolean>
+
+export default TVerifyProofs