+++ src/react-entry.ts
@@ -1,0 +1,3 @@
+  import BringIDModal from './modules/bring-id-modal'
+
+  export { BringIDModal }