+++ src/types/call3.tsx
@@ -1,0 +1,7 @@
+type ICall3 = {
+  target: string
+  allowFailure: boolean
+  callData: string
+}
+
+export default ICall3