+++ src/utils/index.ts
@@ -5,14 +5,18 @@ import defineApiUrl from './define-api-url';
 import api from './api';
 import createQueryString from './create-query-string';
 import msToTime from './ms-to-time';
-import defineExplorerURL from './define-explorer-url';
+import isValidInboundMessage from './validate-inbound-message';
+import isValidOutboundMessage from './validate-outbound-message';
+import fetchRegistryConfig from './fetch-registry-config'
 
 export {
   generateId,
   shortenString,
-  defineExplorerURL,
   msToTime,
   defineApiUrl,
   api,
-  createQueryString
+  createQueryString,
+  isValidInboundMessage,
+  isValidOutboundMessage,
+  fetchRegistryConfig
 }